tools_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile tools_requirements.in
#
certifi==2019.11.28       # via httpx
chardet==3.0.4            # via httpx
gitdb2==2.0.6             # via gitpython
gitpython==3.0.5
h11==0.8.1                # via httpx
h2==3.1.1                 # via httpx
hpack==3.0.0              # via h2
hstspreload==2019.12.6    # via httpx
httpx==0.9.3
hyperframe==5.2.0         # via h2
idna==2.8                 # via httpx
rfc3986==1.3.2            # via httpx
smmap2==2.0.5             # via gitdb2
sniffio==1.1.0            # via httpx
tabulate==0.8.6
